backend/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
pydantic
openai
python-dotenv
alembic
passlib[bcrypt]
python-multipart
pytest
pytest-asyncio
httpx
 ruff
black
